исправил EffectFadiIn и Player
